Revert "Merge pull request #1 from josephworks/master"
This reverts commit f753d46a32b08f2f6588ec2440024824b0dd0599, reversing
changes made to 52b6861b83eb1d0fa308d1cb234a9d7acb67e19b.

+++ KingGenDesktop/AltSaver.cs
@@ -1,4 +1,4 @@
-﻿using KingGen.Models;
+﻿using KingGen.Net.Models;
 using Newtonsoft.Json;//Literally the best library ever
 using System;
 using System.Collections.Generic;
@@ -6,42 +6,53 @@ using System.IO;
 using System.Windows.Forms;
 // ReSharper disable AssignNullToNotNullAttribute
 
-namespace KingGenDesktop;
-
-internal static class AltSaver
+namespace KingGenDesktop
 {
-    public static void Save()
+    internal static class AltSaver
     {
-        try
+        public static void Save()
         {
-            if (MainForm.GeneratedAlts == null) return;
-            var savedAlts = JsonConvert.DeserializeObject<List<Alt>>(File.ReadAllText(@"alts.json"));
-            savedAlts!.AddRange(MainForm.GeneratedAlts);
-            File.AppendAllText("alts.json", JsonConvert.SerializeObject(savedAlts));
+            try
+            {
+                if (MainForm.GeneratedAlts != null)
+                {
+                    List<Alt> savedAlts = JsonConvert.DeserializeObject<List<Alt>>(File.ReadAllText(@"alts.json"));
+                    if (savedAlts != null)//If the alts in the file are not null then we append the file
+                    {
+                        savedAlts.AddRange(MainForm.GeneratedAlts);//Add all of the newely generated alts to the old list to create a combined list
+                        File.WriteAllText("alts.json", JsonConvert.SerializeObject(savedAlts)); //Save said alts
+                    }
+                    else if (MainForm.GeneratedAlts.Count != 0)
+                    {
+                        File.WriteAllText("alts.json", JsonConvert.SerializeObject(MainForm.GeneratedAlts));
+                    }
+                }
+            }
+            catch (Exception ex)
+            {
+                MessageBox.Show(ex.Message);
+                Console.WriteLine($@"Error: {ex} | Contact Flash_5420 if this issue persists");
+            }
         }
-        catch (Exception ex)
-        {
-            Console.WriteLine($@"Error: {ex} | Contact Flash_5420 if this issue persists");
-        }
-    }
 
-    public static void Load()
-    {
-        try
+        public static void Load()
         {
-            if (File.Exists("alts.json") && File.ReadAllText("alts.json") != "")
+            try
             {
-                MainForm.GeneratedAlts.AddRange(JsonConvert.DeserializeObject<List<Alt>>(File.ReadAllText(@"alts.json")));//Add the deserialized object list to the already generated list
+                if (File.Exists("alts.json") && File.ReadAllText("alts.json") != "")
+                {
+                    MainForm.GeneratedAlts.AddRange(JsonConvert.DeserializeObject<List<Alt>>(File.ReadAllText(@"alts.json")));//Add the deserialized object list to the already generated list
+                }
+                else
+                {
+                    if (!File.Exists("alts.json"))
+                        File.Create("alts.json");
+                }
             }
-            else
+            catch (Exception ex)
             {
-                File.Create("alts.json");
-                MessageBox.Show(@"No alts.json file found in the executing directory");
+                Console.WriteLine($@"Error: {ex} | Contact Flash_5420 if this issue persists");
             }
-        }
-        catch (Exception ex)
-        {
-            Console.WriteLine($@"Error: {ex} | Contact Flash_5420 if this issue persists");
         }
     }
 }
